+++ src/dipdup/config.py
@@ -46,6 +46,7 @@ from dipdup.datasources.subscription import OriginationSubscription
 from dipdup.datasources.subscription import Subscription
 from dipdup.datasources.subscription import TokenTransferSubscription
 from dipdup.datasources.subscription import TransactionSubscription
+from dipdup.enums import LoggingValues
 from dipdup.enums import OperationType
 from dipdup.enums import ReindexingAction
 from dipdup.enums import ReindexingReason
@@ -1217,6 +1218,7 @@ class DipDupConfig:
     prometheus: Optional[PrometheusConfig] = None
     advanced: AdvancedConfig = AdvancedConfig()
     custom: Dict[str, Any] = field(default_factory=dict)
+    logging: LoggingValues = LoggingValues.default
 
     def __post_init_post_parse__(self):
         self.paths: List[str] = []
@@ -1339,6 +1341,16 @@ class DipDupConfig:
         if not isinstance(datasource, TzktDatasourceConfig):
             raise ConfigurationError('`datasource` field must refer to TzKT datasource')
         return datasource
+
+    def set_up_logging(self) -> None:
+        if self.logging == LoggingValues.default:
+            pass
+        elif self.logging == LoggingValues.quiet:
+            logging.getLogger('dipdup').setLevel(logging.WARNING)
+        elif self.logging == LoggingValues.verbose:
+            logging.getLogger('dipdup').setLevel(logging.DEBUG)
+        else:
+            raise RuntimeError(f'Unknown `logging` field value: `{self.logging}`')
 
     def _import_index(self, index_config: IndexConfigT) -> None:
         _logger.debug('Loading callbacks and typeclasses of index `%s`', index_config.name)